small changes for debug/testing

+++ src/main/java/dmacc/repository/AuditoriumRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import dmacc.beans.Auditorium;
 
-public interface AuditoriumRepository extends JpaRepository<Auditorium, Integer> {
+public interface AuditoriumRepository extends JpaRepository<Auditorium, Long> {
 
 }
